either CLI or Cli, but CLi is just awkward
