+++ Thessia
@@ -14,7 +14,7 @@ require_once(__DIR__ . "/init.php");
 // Launch the CLI
 $cli = new Symfony\Component\Console\Application("Embrace Eternity");
 
-$tasks = array(__DIR__ . "/tasks/*.php", __DIR__ . "/tasks/CLi/*.php");
+$tasks = array(__DIR__ . "/tasks/*.php", __DIR__ . "/tasks/CLI/*.php");
 // Load all the commands
 
 foreach ($tasks as $task) {
