+++ tasks/CLI/CreateHelper.php
@@ -23,29 +23,29 @@
  * SOFTWARE.
  */
 
-namespace Thessia\Tasks\CLi;
+namespace Thessia\Tasks\CLI;
 
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Question\ConfirmationQuestion;
 
-class CreateLibrary extends Command
+class CreateHelper extends Command
 {
     protected function configure()
     {
         $this
-            ->setName("create:library")
-            ->setDescription("Create a library file");
+            ->setName("create:helper")
+            ->setDescription("Create a helper");
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        $name = prompt("Name of Library");
+        $name = prompt("Name of Helper");
 
-        $directory = __DIR__ . "/../src/Lib/";
+        $directory = __DIR__ . "/../src/Helpers/";
 
-        $file = file_get_contents(__DIR__ . "/../scaffolds/library_template.txt");
+        $file = file_get_contents(__DIR__ . "/../scaffolds/helper_template.txt");
 
         $file = str_replace("?name", ucfirst($name), $file);
 
@@ -74,9 +74,9 @@ class CreateLibrary extends Command
             fwrite($fh, $file);
             fclose($fh);
             $className = ucfirst($name) . ".php";
-            $output->writeln("Success, Library {$name} has been created");
+            $output->writeln("Success, {$name} has been created");
         } else {
-            $output->writeln("Error, Library {$name} already exists.");
+            $output->writeln("Error, {$name} already exists.");
         }
     }
 }
