+++ tasks/CLI/PhinxInit.php
@@ -23,28 +23,21 @@
  * SOFTWARE.
  */
 
-namespace Thessia\Tasks\CLi;
+namespace Thessia\Tasks\CLI;
 
-use Symfony\Component\Console\Input\InputOption;
+use Symfony\Component\Console\Input\InputArgument;
 
-class PhinxTest extends \Phinx\Console\Command\Test
+class PhinxInit extends \Phinx\Console\Command\Init
 {
     protected function configure()
     {
-        parent::configure();
-
-        $this->addOption('--environment', '-e', InputOption::VALUE_REQUIRED, 'The target environment');
-
-        $this->setName('phinx:test')
-            ->setDescription('Verify the configuration file')
-            ->setHelp(
-                <<<EOT
-                The <info>test</info> command verifies the YAML configuration file and optionally an environment
-
-<info>phinx test</info>
-<info>phinx test -e development</info>
-
-EOT
-            );
+        $this->setName('phinx:init')
+            ->setDescription('Initialize the application for Phinx')
+            ->addArgument('path', InputArgument::OPTIONAL, 'Which path should we initialize for Phinx?')
+            ->setHelp(sprintf(
+                '%sInitializes the application for Phinx%s',
+                PHP_EOL,
+                PHP_EOL
+            ));
     }
 }
