+++ tasks/CLI/PhinxRollback.php
@@ -23,28 +23,31 @@
  * SOFTWARE.
  */
 
-namespace Thessia\Tasks\CLi;
+namespace Thessia\Tasks\CLI;
 
 use Symfony\Component\Console\Input\InputOption;
 
-class PhinxStatus extends \Phinx\Console\Command\Status
+class PhinxRollback extends \Phinx\Console\Command\Rollback
 {
     protected function configure()
     {
         parent::configure();
 
-        $this->addOption('--environment', '-e', InputOption::VALUE_REQUIRED, 'The target environment.');
+        $this->addOption('--environment', '-e', InputOption::VALUE_REQUIRED, 'The target environment');
 
-        $this->setName('phinx:status')
-            ->setDescription('Show migration status')
-            ->addOption('--format', '-f', InputOption::VALUE_REQUIRED,
-                'The output format: text or json. Defaults to text.')
+        $this->setName('phinx:rollback')
+            ->setDescription('Rollback the last or to a specific migration')
+            ->addOption('--target', '-t', InputOption::VALUE_REQUIRED, 'The version number to rollback to')
+            ->addOption('--date', '-d', InputOption::VALUE_REQUIRED, 'The date to rollback to')
             ->setHelp(
                 <<<EOT
-                The <info>status</info> command prints a list of all Migrations, along with their current status
+                The <info>rollback</info> command reverts the last migration, or optionally up to a specific version
 
-<info>phinx status -e development</info>
-<info>phinx status -e development -f json</info>
+<info>phinx rollback -e development</info>
+<info>phinx rollback -e development -t 20111018185412</info>
+<info>phinx rollback -e development -d 20111018</info>
+<info>phinx rollback -e development -v</info>
+
 EOT
             );
     }
