+++ tasks/CLI/PhinxSeedCreate.php
@@ -23,32 +23,23 @@
  * SOFTWARE.
  */
 
-namespace Thessia\Tasks\CLi;
+namespace Thessia\Tasks\CLI;
 
-use Symfony\Component\Console\Input\InputOption;
+use Symfony\Component\Console\Input\InputArgument;
 
-class PhinxRollback extends \Phinx\Console\Command\Rollback
+class PhinxSeedCreate extends \Phinx\Console\Command\SeedCreate
 {
     protected function configure()
     {
         parent::configure();
 
-        $this->addOption('--environment', '-e', InputOption::VALUE_REQUIRED, 'The target environment');
-
-        $this->setName('phinx:rollback')
-            ->setDescription('Rollback the last or to a specific migration')
-            ->addOption('--target', '-t', InputOption::VALUE_REQUIRED, 'The version number to rollback to')
-            ->addOption('--date', '-d', InputOption::VALUE_REQUIRED, 'The date to rollback to')
-            ->setHelp(
-                <<<EOT
-                The <info>rollback</info> command reverts the last migration, or optionally up to a specific version
-
-<info>phinx rollback -e development</info>
-<info>phinx rollback -e development -t 20111018185412</info>
-<info>phinx rollback -e development -d 20111018</info>
-<info>phinx rollback -e development -v</info>
-
-EOT
-            );
+        $this->setName('phinx:seedcreate')
+            ->setDescription('Create a new database seeder')
+            ->addArgument('seedname', InputArgument::REQUIRED, 'What is the name of the seeder?')
+            ->setHelp(sprintf(
+                '%sCreates a new database seeder%s',
+                PHP_EOL,
+                PHP_EOL
+            ));
     }
 }
