+++ tasks/CLI/PhinxStatus.php
@@ -23,28 +23,28 @@
  * SOFTWARE.
  */
 
-namespace Thessia\Tasks\CLi;
+namespace Thessia\Tasks\CLI;
 
 use Symfony\Component\Console\Input\InputOption;
 
-class PhinxSeedRun extends \Phinx\Console\Command\SeedRun
+class PhinxStatus extends \Phinx\Console\Command\Status
 {
     protected function configure()
     {
         parent::configure();
 
-        $this->addOption('--environment', '-e', InputOption::VALUE_REQUIRED, 'The target environment');
+        $this->addOption('--environment', '-e', InputOption::VALUE_REQUIRED, 'The target environment.');
 
-        $this->setName('phinx:seedrun')
-            ->setDescription('Run database seeders')
+        $this->setName('phinx:status')
+            ->setDescription('Show migration status')
+            ->addOption('--format', '-f', InputOption::VALUE_REQUIRED,
+                'The output format: text or json. Defaults to text.')
             ->setHelp(
                 <<<EOT
-                The <info>seed:run</info> command runs all available or individual seeders
+                The <info>status</info> command prints a list of all Migrations, along with their current status
 
-<info>phinx seed:run -e development</info>
-<info>phinx seed:run -e development -s UserSeeder</info>
-<info>phinx seed:run -e development -v</info>
-
+<info>phinx status -e development</info>
+<info>phinx status -e development -f json</info>
 EOT
             );
     }
